requirement.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-import-export==2.4.0
django-simpleui==2020.9.26
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Admin==1.5.7
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
pycodestyle==2.6.0
pylint==2.6.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xlrd==1.2.0
xlwt==1.3.0
